Fix check for inf2cat and signtool existence
Script fails when it doesn't find a match in $inf2cat_paths or $signtool_paths.  Test-Path fails when passed a null value and that failure does not translate to $false, so -not cannot make it $true.  If the first check on a null value is [String]::IsNullOrEmpty, Test-Path won't run and won't throw an error.